Use --show edit to search in change descriptions
A note about ebb merge:
if the parent branch was already merged, that parent branch cannot be edited. So when merge attempts to update files it creates a new branch which will continue being used as parent for other child branches. Its parent is the merged branch grandparent.
It is actually intuitive in use.