Add --sync merge for a more traditional workflow
By default, ebb encourages to create task branches where the parent is frozen. This is supported by the commands freeze and merge without options.
For traditional workflows where the parent branch might change concurrently (local or remote changes), it is recommended to sync these changes to the child branch regularly.